add notebook to explore edm-recipes data catalog
